+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- siteConfigs query
+- storeConfigs query
 
 ## [2.90.8] - 2019-07-09
 ### Changed